DEXW-4022: Fix types and clean
