+++ nginx/webkeeper.conf
@@ -7,8 +7,9 @@ server {
     add_header X-Content-Type-Options nosniff always;
     add_header X-XSS-Protection "1; mode=block" always;
     add_header strict-transport-security "max-age=2592000; includeSubDomains" always;
-        location / {
-            root /iframe-entry/;
-        }
+
+    location / {
+        root /iframe-entry/;
+    }
 }
