+++ public
@@ -1,1 +1,1 @@
-src/iframe-entry/public
+./src/iframe-entry/public
