+++ src/iframe-entry/components/InvokeFunction/InvokeFunction.tsx
@@ -22,7 +22,10 @@ export const InvokeFunction: FC<IProps & TFlexProps> = ({
             <Text sx={pseudoElemStyles}>{name}</Text>
             <Text sx={wrapperStylesStart}>
                 {args.map(({ type, value }, index) => (
-                    <Text sx={getAttrStyles(type, index === args.length - 1)}>
+                    <Text
+                        key={index}
+                        sx={getAttrStyles(type, index === args.length - 1)}
+                    >
                         {getAttrContent(type, value)}
                     </Text>
                 ))}
