+++ src/iframe-entry/handlers/sign.ts
@@ -1,12 +1,12 @@
 import { SignedTx, SignerTx } from '@waves/signer';
 import { IUser } from '../../interface';
-import { Queue } from '../../utils/Queue';
+import { IQueue } from '../../utils/Queue';
 import { IState } from '../interface';
 import sign from '../router/sign';
 import { loadUserData, preload, toQueue } from './helpers';
 
 export const getSignHandler = (
-    queue: Queue,
+    queue: IQueue,
     state: IState
 ): ((list: Array<SignerTx>) => Promise<Array<SignedTx<SignerTx>>>) =>
     toQueue(queue, (list: Array<SignerTx>) => {
