+++ src/iframe-entry/handlers/signMessage.ts
@@ -1,11 +1,11 @@
 import { IUser } from '../../interface';
-import { Queue } from '../../utils/Queue';
+import { IQueue } from '../../utils/Queue';
 import { IState } from '../interface';
 import signMessage from '../router/signMessage';
 import { loadUserData, preload, toQueue } from './helpers';
 
 export const getSignMessageHandler = (
-    queue: Queue,
+    queue: IQueue,
     state: IState
 ): ((message: string | number) => Promise<string>) =>
     toQueue(queue, (message: string | number) => {
