+++ src/iframe-entry/handlers/signTypedData.ts
@@ -1,12 +1,12 @@
 import { TDataEntry } from '@waves/waves-transactions';
 import { IUser } from '../../interface';
-import { Queue } from '../../utils/Queue';
+import { IQueue } from '../../utils/Queue';
 import { IState } from '../interface';
 import signTypedData from '../router/signTypedData';
 import { loadUserData, preload, toQueue } from './helpers';
 
 export const getSignTypedDataHandler = (
-    queue: Queue,
+    queue: IQueue,
     state: IState
 ): ((data: Array<TDataEntry>) => Promise<string>) =>
     toQueue(queue, (data: Array<TDataEntry>) => {
