+++ src/iframe-entry/pages/SignTransfer/__tests__/helpers.test.ts
@@ -1,13 +1,13 @@
-import * as helpers from '../helpers';
-import { WAVES, NAME_MAP } from '../../../constants';
+import { AssetDecimals, Long, MassTransferItem } from '@waves/ts-types';
+import { NAME_MAP, WAVES } from '../../../constants';
 import * as math from '../../../utils/math';
-import { MassTransferItem, Long } from '@waves/ts-types';
+import * as helpers from '../helpers';
 
 const massTransferTxType = NAME_MAP.massTransfer;
 const transferTxType = NAME_MAP.transfer;
 const asset1Id = 'asset1';
 const asset1Name = 'asset1_name';
-const asset1Decimals = 2;
+const asset1Decimals = 2 as AssetDecimals;
 const asset1 = {
     assetId: asset1Id,
     decimals: asset1Decimals,
