+++ src/iframe-entry/router/batch.ts
@@ -1,10 +1,9 @@
 import { SignerTx } from '@waves/signer';
 import { Transaction } from '@waves/ts-types';
 import { libs, signTx } from '@waves/waves-transactions';
-import { IUser } from '../../interface';
+import { ITransactionInfo, IUser } from '../../interface';
 import { IState } from '../interface';
 import batchPage from '../pages/batch';
-import { ITransactionInfo } from '../services/transactionsService';
 import renderPage from '../utils/renderPage';
 
 export default function (
