+++ src/iframe-entry/utils/loadLogoInfo.ts
@@ -1,7 +1,8 @@
-import { TAssetDetails } from '@waves/node-api-js/es/api-node/assets';
 import { fetchDataKey } from '@waves/node-api-js/es/api-node/addresses';
-import curry from 'ramda/es/curry';
+import { TAssetDetails } from '@waves/node-api-js/es/api-node/assets';
 import { DataTransactionEntryString, Long } from '@waves/ts-types';
+import curry from 'ramda/es/curry';
+import { DetailsWithLogo } from '../../interface';
 
 const BETTER_TOKENS_MAP = {
     W: '3P6t5mKGwVDkyjFhtUqw4NnecyC3DRpLfkw',
@@ -129,7 +130,3 @@ export const loadLogoInfo = curry(
             })
         )
 );
-
-export type DetailsWithLogo = TAssetDetails<Long> & {
-    logo?: string;
-};
